Extract body structure view
