+++ crates/engine/ir-view/src/body/mod.rs
@@ -1,0 +1,7 @@
+//! Source-shape projections derived from lowered Body IR.
+
+mod structure;
+
+pub use structure::{
+    BodyClosingBraceBlock, BodyClosingBraceBlockKind, BodyStructureView, MethodChainExprTy,
+};
