+++ crates/engine/ir-view/src/ty/locals.rs
@@ -3,13 +3,13 @@
 //! Body IR owns lowered expression, scope, and local declaration storage. This view exposes the
 //! parts that higher analysis features need without making them know the Body IR query vocabulary.
 
-use std::collections::{HashMap, HashSet};
+use std::collections::HashSet;
 
 use rg_body_ir::BindingKind;
 use rg_ir_model::{
-    BindingId, BodyBindingRef, BodyRef, DefMapRef, ExprId, ExprKind, ExprWrapperKind, FunctionRef,
-    ModuleId, ModuleRef, ScopeId, SemanticItemKind, SemanticItemRef, TargetRef,
-    hir::source::ItemSourceKind, identity::DeclarationRef,
+    BindingId, BodyBindingRef, BodyRef, DefMapRef, ExprId, ExprKind, FunctionRef, ModuleId,
+    ModuleRef, ScopeId, SemanticItemKind, SemanticItemRef, TargetRef, hir::source::ItemSourceKind,
+    identity::DeclarationRef,
 };
 use rg_ir_storage::ItemStoreQuery;
 use rg_parse::{FileId, Span, TextSpan};
@@ -134,60 +134,6 @@ impl ResolvedFunctionCall {
 
     pub fn args(&self) -> &[ResolvedCallArg] {
         &self.args
-    }
-}
-
-/// A body-derived construct whose known source span can be labeled after its closing brace.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct BodyClosingBraceBlock {
-    file_id: FileId,
-    span: Span,
-    label: String,
-    label_span: Option<Span>,
-}
-
-impl BodyClosingBraceBlock {
-    pub fn file_id(&self) -> FileId {
-        self.file_id
-    }
-
-    pub fn span(&self) -> Span {
-        self.span
-    }
-
-    pub fn label(&self) -> &str {
-        &self.label
-    }
-
-    pub fn label_span(&self) -> Option<Span> {
-        self.label_span
-    }
-}
-
-/// A typed method call that feeds another method segment in a chain.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct MethodChainExprTy {
-    file_id: FileId,
-    span: Span,
-    parent_dot_span: Span,
-    ty: Ty,
-}
-
-impl MethodChainExprTy {
-    pub fn file_id(&self) -> FileId {
-        self.file_id
-    }
-
-    pub fn span(&self) -> Span {
-        self.span
-    }
-
-    pub fn parent_dot_span(&self) -> Span {
-        self.parent_dot_span
-    }
-
-    pub fn ty(&self) -> &Ty {
-        &self.ty
     }
 }
 
@@ -553,102 +499,6 @@ impl<'a, 'db> BodyView<'a, 'db> {
         Ok(calls)
     }
 
-    /// Returns known expression types for intermediate segments in method chains.
-    ///
-    /// The projection stays chain-shaped instead of exposing every typed expression. That keeps
-    /// early expression type hints close to the high-signal rust-analyzer default shape.
-    pub fn method_chain_expr_tys(
-        &self,
-        target: TargetRef,
-        file_id: FileId,
-    ) -> anyhow::Result<Vec<MethodChainExprTy>> {
-        let Some(target_bodies) = self.db.body_ir.target_bodies(target)? else {
-            return Ok(Vec::new());
-        };
-
-        let mut expr_tys = Vec::new();
-        for body in target_bodies.bodies() {
-            let parent_dot_by_receiver = Self::method_parent_dots_by_receiver(body);
-
-            for (expr_idx, expr) in body.exprs().iter().enumerate() {
-                if expr.source.file_id != file_id {
-                    continue;
-                }
-                if !matches!(expr.kind, ExprKind::MethodCall { .. }) {
-                    continue;
-                }
-                let Some(parent_dot_span) = parent_dot_by_receiver.get(&ExprId(expr_idx)).copied()
-                else {
-                    continue;
-                };
-                let Some(ty) = body.expr_ty(ExprId(expr_idx)).cloned() else {
-                    continue;
-                };
-                if matches!(ty, Ty::Unknown) {
-                    continue;
-                }
-
-                expr_tys.push(MethodChainExprTy {
-                    file_id: expr.source.file_id,
-                    span: expr.source.span,
-                    parent_dot_span,
-                    ty,
-                });
-            }
-        }
-
-        Ok(expr_tys)
-    }
-
-    /// Returns body-owned blocks whose source extent ends at their closing brace.
-    ///
-    /// Closing-brace hints only need the block-level source span, not the body syntax tree that
-    /// originally produced it. This keeps the analysis path tied to durable Body IR facts.
-    pub fn closing_brace_blocks(
-        &self,
-        target: TargetRef,
-        file_id: FileId,
-    ) -> anyhow::Result<Vec<BodyClosingBraceBlock>> {
-        let Some(target_bodies) = self.db.body_ir.target_bodies(target)? else {
-            return Ok(Vec::new());
-        };
-
-        let items = ItemStoreQuery::new(self.db);
-        let mut blocks = Vec::new();
-        for body in target_bodies.bodies() {
-            let body_source = body.source();
-            if body_source.file_id == file_id
-                && let Some(function) = body.function_owner()
-                && let Some(function) = items.function_data(function)?
-            {
-                blocks.push(BodyClosingBraceBlock {
-                    file_id: body_source.file_id,
-                    span: body_source.span,
-                    label: format!("// fn {}", function.name),
-                    label_span: None,
-                });
-            }
-
-            for expr in body.exprs() {
-                if expr.source.file_id != file_id {
-                    continue;
-                }
-                let Some(label) = Self::closing_brace_label(&expr.kind) else {
-                    continue;
-                };
-
-                blocks.push(BodyClosingBraceBlock {
-                    file_id: expr.source.file_id,
-                    span: expr.source.span,
-                    label: label.to_string(),
-                    label_span: Self::closing_brace_label_span(body, &expr.kind),
-                });
-            }
-        }
-
-        Ok(blocks)
-    }
-
     pub fn local_groups(
         &self,
         target: TargetRef,
@@ -760,66 +610,5 @@ impl<'a, 'db> BodyView<'a, 'db> {
                 })
             })
             .collect()
-    }
-
-    fn method_parent_dots_by_receiver(
-        body: &rg_body_ir::ResolvedBodyData,
-    ) -> HashMap<ExprId, Span> {
-        let mut parent_dot_by_receiver = HashMap::new();
-        for expr in body.exprs() {
-            let ExprKind::MethodCall {
-                receiver: Some(receiver),
-                dot_span: Some(dot_span),
-                ..
-            } = &expr.kind
-            else {
-                continue;
-            };
-            let receiver = Self::chain_receiver_base(body, *receiver);
-            parent_dot_by_receiver.entry(receiver).or_insert(*dot_span);
-        }
-
-        parent_dot_by_receiver
-    }
-
-    fn chain_receiver_base(body: &rg_body_ir::ResolvedBodyData, receiver: ExprId) -> ExprId {
-        let mut current = receiver;
-        while let Some(expr) = body.expr(current) {
-            let ExprKind::Wrapper {
-                kind: ExprWrapperKind::Paren | ExprWrapperKind::Try | ExprWrapperKind::Await,
-                inner: Some(inner),
-            } = &expr.kind
-            else {
-                break;
-            };
-            current = *inner;
-        }
-
-        current
-    }
-
-    fn closing_brace_label(expr: &ExprKind) -> Option<&'static str> {
-        match expr {
-            ExprKind::Match { .. } => Some("// match"),
-            ExprKind::Loop { .. } => Some("// loop"),
-            ExprKind::While { .. } => Some("// while"),
-            ExprKind::For { .. } => Some("// for"),
-            _ => None,
-        }
-    }
-
-    fn closing_brace_label_span(
-        body: &rg_body_ir::ResolvedBodyData,
-        expr: &ExprKind,
-    ) -> Option<Span> {
-        let ExprKind::Match {
-            scrutinee: Some(scrutinee),
-            ..
-        } = expr
-        else {
-            return None;
-        };
-
-        body.expr(*scrutinee).map(|expr| expr.source.span)
     }
 }
